Refactor login screen to use a view model
