+++ lib/main.dart
@@ -1,4 +1,4 @@
-import 'package:financeplanner/views/login_screen.dart';
+import 'package:financeplanner/views/login_view/login_screen.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_localizations/flutter_localizations.dart';
